requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async SQLAlchemy and PostgreSQL driver
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0

# Pydantic v2 with settings management
pydantic==2.5.3
pydantic-settings==2.1.0

# Python standard library enhancements
python-dotenv==1.0.0

# HTTP client for async API calls and example scripts
httpx==0.27.0
requests==2.31.0

# AWS S3 client
aioboto3==12.3.0
